refactor: 유틸 함수 폴더를 domain과 view로 세분화
