+++ src/utils/domain/predicate.js
