+++ src/utils/domain/predicate.test.js
