+++ src/validation/validateBonusNumber.js
@@ -1,25 +1,34 @@
 import CustomError from "../CustomError.js";
 import { ERROR_MESSAGE, ERROR_PREFIX } from "../constants/message.js";
 import { LOTTO_RULE } from "../constants/lotto.js";
-import { isInRange, isDuplicate } from "../utils/predicate.js";
+import { isInRange, isDuplicate } from "../utils/domain/predicate.js";
 
 const validateBonusNumber = (bonusNumber, winningNumbers) => {
   if (!Number.isSafeInteger(bonusNumber)) {
-    throw new CustomError(ERROR_MESSAGE.INVALID_INTEGER, ERROR_PREFIX.typeError);
+    throw new CustomError(
+      ERROR_MESSAGE.INVALID_INTEGER,
+      ERROR_PREFIX.typeError,
+    );
   }
 
   if (
     !isInRange(
       bonusNumber,
       LOTTO_RULE.MIN_LOTTO_NUMBER,
-      LOTTO_RULE.MAX_LOTTO_NUMBER
+      LOTTO_RULE.MAX_LOTTO_NUMBER,
     )
   ) {
-    throw new CustomError(ERROR_MESSAGE.INVALID_LOTTO_NUMBER_RANGE, ERROR_PREFIX.rangeError);
+    throw new CustomError(
+      ERROR_MESSAGE.INVALID_LOTTO_NUMBER_RANGE,
+      ERROR_PREFIX.rangeError,
+    );
   }
 
   if (isDuplicate([bonusNumber, ...winningNumbers])) {
-    throw new CustomError(ERROR_MESSAGE.INVALID_DUPLICATE_BONUS_NUMBER, ERROR_PREFIX.duplicateError);
+    throw new CustomError(
+      ERROR_MESSAGE.INVALID_DUPLICATE_BONUS_NUMBER,
+      ERROR_PREFIX.duplicateError,
+    );
   }
 };
